Move radar to Render category
